Refactoring to allow more control of encoding pattern in transports that have a more specific encoding.

+++ CCSDS_MAL_ENCODING_BINARY_FIXED/src/main/java/esa/mo/mal/encoder/binary/fixed/FixedBinaryDecoder.java
@@ -119,7 +119,7 @@ public class FixedBinaryDecoder extends esa.mo.mal.encoder.binary.BinaryDecoder
     {
       checkBuffer(2);
       final int i = shiftOffsetAndReturnPrevious(2);
-      return java.nio.ByteBuffer.wrap(getBuf(), i, 2).getInt() & 0xFFFF;
+      return java.nio.ByteBuffer.wrap(getBuf(), i, 2).getShort() & 0xFFFF;
     }
 
     @Override
